+++ pbge/namegen.py
@@ -1,285 +1,143 @@
-"""
-Copyright 2010 Joao Henriques <jotaf (no spam) at hotmail dot com>.
-
-Modified by Joseph Hewitt for Dungeon Monkey Eternal.
-
-This file is part of name-gen.
-
-name-gen is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as
-published by the Free Software Foundation, either version 3 of the
-License, or (at your option) any later version.
-
-name-gen is distributed in the hope that it will be useful, but
-WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public
-License along with name-gen.  If not, see
-<http://www.gnu.org/licenses/>.
-"""
-
-import itertools
-import random
-import locale
-from . import util
-import collections
 import json
+import random
+
+class Grap:
+    # I was going to name this class Grapheme, but it's not necessarily
+    # a grapheme, so I just named it Grap.
+    def __init__(self, next_syllables):
+        self.next_syllables = next_syllables
+
+    def get_next_candidates(self):
+        candidates = list()
+        for k,v in self.next_syllables.items():
+            candidates += [k,] * v
+        return candidates
+        
+    def gen_next_syllable(self):
+        candidates = self.get_next_candidates()
+        if candidates:
+            return random.choice(candidates)
+        else:
+            return " "
+
+
+class Syllable(Grap):
+    def __init__(self, next_syllables, nuco_prime, nuco_secondary, onset):
+        super().__init__(next_syllables)
+        self.nuco_prime = nuco_prime
+        self.nuco_secondary = nuco_secondary
+        self.onset = onset
+
 
 class NameGen:
-	"""
-	name-gen: Free python name generator module that analyzes sample text and produces
-	similar words.
-	
-	Usage:
-		1. Initialize with path to language file (generated using 'namegen_training.py').
-		2. Call gen_word() method, returns generated string.
-	
-	Optional:
-		. Change min_syl and max_syl to control number of syllables.
-		. Pass the sample file as 2nd parameter at initialization to set it as the list
-		  of forbidden words. No words from the sample will be replicated.
-		. Pass True as the 1st parameter to name_gen() to add the generated word to the
-		  list of forbidden words. The word will not occur again.
-	"""
-	
-	def __init__(self, language_file, forbidden_file = None, min_syl=2, max_syl=4):
-		self.min_syl = min_syl
-		self.max_syl = max_syl
-		
-		#load language file
-		with open(util.data_dir(language_file), 'r') as f:
-			lines = [line.strip() for line in f.readlines()]
-			
-			self.syllables = lines[0].split(',')  #first line, list of syllables
-			
-			starts_ids = [int(n) for n in lines[1].split(',')]  #next 2 lines, start syllable indexes and counts
-			starts_counts = [int(n) for n in lines[2].split(',')]
-			self.starts = list(zip(starts_ids, starts_counts))  #zip into a list of tuples
-			
-			ends_ids = [int(n) for n in lines[3].split(',')]  #next 2, same for syllable ends
-			ends_counts = [int(n) for n in lines[4].split(',')]
-			self.ends = list(zip(ends_ids, ends_counts))
-			
-			#starting with the 6th and 7th lines, each pair of lines holds ids and counts
-			#of the "next syllables" for a previous syllable.
-			self.combinations = []
-			for (ids_str, counts_str) in zip(lines[5:None:2], lines[6:None:2]):
-				if len(ids_str) == 0 or len(counts_str) == 0:  #empty lines
-					self.combinations.append([])
-				else:
-					line_ids = [int(n) for n in ids_str.split(',')]
-					line_counts = [int(n) for n in counts_str.split(',')]
-					self.combinations.append(list(zip(line_ids, line_counts)))
-		
-		#load forbidden words file if needed
-		if forbidden_file is None:
-			self.forbidden = ''
-		else:
-			self.forbidden = _load_sample(forbidden_file)
-	
-	def gen_word(self, no_repeat = True):
-		#random number of syllables, the last one is always appended
-		num_syl = random.randint(self.min_syl, self.max_syl - 1)
-		
-		#turn ends list of tuples into a dictionary
-		ends_dict = dict(self.ends)
-		
-		#we may have to repeat the process if the first "min_syl" syllables were a bad choice
-		#and have no possible continuations; or if the word is in the forbidden list.
-		word = []; word_str = ''
-		while len(word) < self.min_syl or self.forbidden.find(word_str) != -1:
-			#start word with the first syllable
-			syl = _select_syllable(self.starts, 0)
-			word = [self.syllables[syl]]
-			
-			for i in range(1, num_syl):
-				#don't end yet if we don't have the minimum number of syllables
-				if i < self.min_syl: end = 0
-				else: end = ends_dict.get(syl, 0)  #probability of ending for this syllable
-				
-				#select next syllable
-				syl = _select_syllable(self.combinations[syl], end)
-				if syl is None: break  #early end for this word, end syllable was chosen
-				
-				word.append(self.syllables[syl])
-				
-			else:  #forcefully add an ending syllable if the loop ended without one
-				syl = _select_syllable(self.ends, 0)
-				word.append(self.syllables[syl])
-			
-			word_str = ''.join(word)
-		
-		#to ensure the word doesn't repeat, add it to the forbidden words
-		if no_repeat: self.forbidden = self.forbidden + '\n' + word_str
-		
-		return word_str.capitalize()
+    def __init__(self, fname, forbidden_names=None):
+        # forbidden_names is a set of names that should be rejected if
+        # generated. It's passed as a param so multiple name generators
+        # can share the same list.
+        with open(fname) as f:
+            my_dict = json.load(f)
+            
+        self.syllables = dict()
+        self.nuco = dict()
+        self.onsets = dict()
+        
+        self.min_length = max(min(my_dict.get("min_length", 4), 6), 3)
+        self.max_length = max(min(my_dict.get("max_length", 8), 12), 8)
 
-def _select_syllable(counts, end_count):
-	if len(counts) == 0: return None  #no elements to choose from
-	
-	#"counts" holds cumulative counts, so take the last element in the list
-	#(and 2nd in that tuple) to get the sum of all counts
-	chosen = random.randint(0, counts[-1][1] + end_count)
-	
-	for (syl, count) in counts:
-		if count >= chosen:
-			return syl
-	return None
+        for k,v in my_dict["syllables"].items():
+            self.syllables[k] = Syllable(**v)
+        
+        for k,v in my_dict["nuco"].items():
+            self.nuco[k] = Grap(**v)
 
-def _load_sample(filename):
-	#get sample text
-	with open(util.data_dir(filename), 'r') as f:
-		sample = ''.join(f.readlines()).lower()
-	
-	#convert accented characters to non-accented characters
-	sample = locale.strxfrm(sample)
-	
-	#remove all characters except letters from A to Z
-	a = ord('a')
-	z = ord('z')
-	sample = ''.join([
-		c if (ord(c) >= a and ord(c) <= z) else ' '
-			for c in sample])
-	
-	return sample
+        for k,v in my_dict["onsets"].items():
+            self.onsets[k] = Grap(**v)
 
-class KoreanNameGen:
-	def __init__(self, language_file, forbidden_file = None, min_syl=2, max_syl=4):
-		with open(util.data_dir(language_file), 'rt') as f:
-			self.syllables = json.load(f)
+        if forbidden_names:
+            self.forbidden_names = forbidden_names
+        else:
+            self.forbidden_names = set()
 
-	NULL = 'null'
-	ALL_SYLLABLES = 'ALL_SYLLABLES'
+    def gen_syllable(self, prev_syllable=" "):
+        if prev_syllable in self.syllables:
+            candidates = self.syllables[prev_syllable].get_next_candidates()
+            onset_candidates = list()
+            for can in candidates:
+                if can in self.syllables and self.syllables[can].onset and self.syllables[can].onset in self.onsets:
+                    #print(self.syllables[can].onset)
+                    onset_candidates += self.onsets[self.syllables[can].onset].get_next_candidates()
+            candidates = candidates * 5 + onset_candidates
+            if self.syllables[prev_syllable].nuco_prime and self.syllables[prev_syllable].nuco_prime in self.nuco:
+                candidates += self.nuco[self.syllables[prev_syllable].nuco_prime].get_next_candidates() * 4
+            if self.syllables[prev_syllable].nuco_secondary and self.syllables[prev_syllable].nuco_secondary in self.nuco:
+                candidates += self.nuco[self.syllables[prev_syllable].nuco_secondary].get_next_candidates() * 3
+            if prev_syllable[-1] in self.nuco:
+                candidates += self.nuco[prev_syllable[-1]].get_next_candidates()
+        else:
+            candidates = ['0']
+        if candidates:
+            return random.choice(candidates)
+        else:
+            return "9"
 
-	def gen_word(self, no_repeat = True):
-		last_glyph = self.NULL
-		my_word = list()
-		while True:
-			candidates = list()
-			if last_glyph in self.syllables:
-				for k,v in self.syllables[last_glyph].items():
-					candidates += [k] * v
-			else:
-				print(last_glyph)
-				candidates = list(self.syllables.keys())
-				candidates.remove('null')
-			next_glyph = random.choice(candidates)
-			if next_glyph != self.NULL:
-				my_word.append(next_glyph)
-				if len(my_word) > random.randint(1,5):
-					break
-			else:
-				break
-		return "".join(my_word)
+    def _gen_word_(self, length=0):
+        proto_name = ""
+        prev_syllable = " "
+        if length < 3:
+            length = (random.randint(3,8) + random.randint(3,8))//2
+            if random.randint(1,23) == 5:
+                length += random.randint(1,6)
+            if random.randint(1,10) == 1:
+                length -= 1
+        increased_length = False
+        while len(proto_name) < length:
+            new_syllable = self.gen_syllable(prev_syllable)
+            if new_syllable == " " and len(proto_name) < random.randint(2,3) and random.randint(1,23) != 5:
+                new_syllable = self.gen_syllable(prev_syllable)
+            if new_syllable == " " and not increased_length:
+                length += random.randint(2,3)
+                increased_length = True
+            proto_name = proto_name + new_syllable
+            prev_syllable = new_syllable
+        proto_name = proto_name.strip()
+        
+        verified_letters = list()
+        needs_capitalization = True
+        prev_char = ""
+        repeat_count = 0
+        for char in proto_name:
+            if char == prev_char:
+                repeat_count += 1
+                if char == " " or repeat_count > 1:
+                    continue
+            else:
+                prev_char = char
+                repeat_count = 0
+            if needs_capitalization:
+                char = char.upper()
+                needs_capitalization = False
+            if char == " ":
+                needs_capitalization = True
+            verified_letters.append(char)
+        return "".join(verified_letters)
 
-	def gen_word2(self, no_repeat = True):
-		a = self.NULL
-		b = self.NULL
-		my_word = list()
-		while True:
-			candidates = list()
-			ab = a + b
-			if ab in self.syllables:
-				for k,v in self.syllables[ab].items():
-					candidates += [k] * v
-			else:
-				candidates = self.syllables[self.ALL_SYLLABLES]
+    def gen_word(self, length=0):
+        myword = ""
+        while not myword:
+            myword = self._gen_word_(length)
+            if myword in self.forbidden_names:
+                myword = ""
+        self.forbidden_names.add(myword)
+        return myword
 
-			next_glyph = random.choice(candidates)
-			if next_glyph != self.NULL:
-				my_word.append(next_glyph)
-				if len(my_word) > random.randint(1,5):
-					break
-			else:
-				break
-		return "".join(my_word)
-
-	@classmethod
-	def generate_library2(cls, training_file, output_file):
-		with open(training_file, "rt") as f:
-			mylines = f.readlines()
-
-		used_names = set()
-		syllables = dict()
-		syllables[cls.ALL_SYLLABLES] = list()
-		for line in mylines:
-			for word in line.split():
-				if len(word) > 1:
-					if word[0] == "(":
-						break
-					elif word not in used_names:
-						used_names.add(word)
-						a = cls.NULL
-						b = cls.NULL
-						for glyph in word:
-							ab = a + b
-							if ab not in syllables:
-								syllables[ab] = collections.defaultdict(int)
-							syllables[ab][glyph] += 1
-							a = b
-							b = glyph
-							syllables[cls.ALL_SYLLABLES].append(glyph)
-						ab = a + b
-						if ab not in syllables:
-							syllables[ab] = collections.defaultdict(int)
-						syllables[ab][cls.NULL] += 1
-
-		with open(output_file, "wt") as f:
-			json.dump(syllables, f)
-
-	@staticmethod
-	def generate_library(training_file, output_file):
-		with open(training_file, "rt") as f:
-			mylines = f.readlines()
-
-		used_names = set()
-		syllables = dict()
-		for line in mylines:
-			for word in line.split():
-				if len(word) > 1:
-					if word[0] == "(":
-						break
-					elif word not in used_names:
-						used_names.add(word)
-						last_glyph = None
-						for glyph in word:
-							if last_glyph not in syllables:
-								syllables[last_glyph] = collections.defaultdict(int)
-							syllables[last_glyph][glyph] += 1
-							last_glyph = glyph
-						if last_glyph not in syllables:
-							syllables[last_glyph] = collections.defaultdict(int)
-						syllables[last_glyph][None] += 1
-
-		with open(output_file, "wt") as f:
-			json.dump(syllables, f)
-
-
-
-def test_names( ng, trials ):
-    tot1 = 0
-    s1 = set()
-
-    mx1 = 0
-    mn1 = 99
-
-    for t in range( trials ):
-        n1 = ng.gen_word()
-        tot1 += len( n1 )
-        mx1 = max( len( n1 ) , mx1 )
-        mn1 = min( len( n1 ) , mn1 )
-        s1.add( n1 )
-
-    print("  average length: " + str( float( tot1 )/trials ))
-    print("  unique names: " + str( len( s1 ) ))
-    print("  max/min: " + str( mx1 ) + "/" + str( mn1 ))
-
-#ALL_NAME_GENERATORS = ( DEFAULT, DWARF, DRAGON, ELF, ORC, GREEK, JAPANESE, GNOME, HURTHLING, ELDRITCH )
-#
-#def random_style_name():
-#    ng = random.choice( ALL_NAME_GENERATORS )
-#    return ng.gen_word()
-
-
+    def test_generator(self):
+        dupes = 0
+        for t in range(10):
+            for _ in range(1000):
+                myname = self._gen_word_()
+                if myname in self.forbidden_names:
+                    dupes += 1
+                self.forbidden_names.add(myname)
+            print((t+1)*1000)
+        print("{}/10000 names duplicated".format(dupes))
+